changed to double solonoid
